refactor(string): create string form component and implement on string add page
